backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
yfinance==0.2.28
pandas==2.1.4
numpy==1.24.3
requests==2.32.4
python-dotenv==1.0.0
anthropic==0.40.0